+++ kondor-core/src/main/kotlin/com/ubertob/kondor/json/JAny.kt
@@ -4,8 +4,6 @@ import com.ubertob.kondor.outcome.Outcome
 import com.ubertob.kondor.outcome.asFailure
 import com.ubertob.kondor.outcome.asSuccess
 import java.util.concurrent.atomic.AtomicReference
-import kotlin.properties.ReadOnlyProperty
-import kotlin.reflect.KProperty
 
 
 typealias NodeWriter<T> = (JsonNodeObject, T) -> JsonNodeObject
@@ -97,43 +95,6 @@ data class JsonPropOptional<T : Any, JN : JsonNode>(override val propName: Strin
                 )
             } ?: wrapped
         }
-
-}
-
-sealed class JFieldBase<T, PT : Any>
-    : ReadOnlyProperty<JAny<PT>, JsonProperty<T>> {
-
-    protected abstract val binder: (PT) -> T
-
-    protected abstract fun buildJsonProperty(property: KProperty<*>): JsonProperty<T>
-
-    operator fun provideDelegate(thisRef: JAny<PT>, prop: KProperty<*>): JFieldBase<T, PT> {
-        val jp = buildJsonProperty(prop)
-        thisRef.registerProperty(jp, binder)
-        return this
-    }
-
-    override fun getValue(thisRef: JAny<PT>, property: KProperty<*>): JsonProperty<T> =
-        buildJsonProperty(property)
-}
-
-class JField<T : Any, PT : Any>(
-    override val binder: (PT) -> T,
-    private val converter: JConverter<T>
-) : JFieldBase<T, PT>() {
-
-    override fun buildJsonProperty(property: KProperty<*>): JsonProperty<T> =
-        JsonPropMandatory(property.name, converter)
-
-}
-
-class JFieldMaybe<T : Any, PT : Any>(
-    override val binder: (PT) -> T?,
-    private val converter: JConverter<T>
-) : JFieldBase<T?, PT>() {
-
-    override fun buildJsonProperty(property: KProperty<*>): JsonProperty<T?> =
-        JsonPropOptional(property.name, converter)
 
 }