+++ scripts/release-on-mavencentral.sh
@@ -1,8 +1,8 @@
 # Steps:
-# verify it's all working with a ./gradlew clean build
 # remove snapshot from the version manually in build.gradle
 # update the README.md
 # check the CHANGELOG.md
+# verify it's all working with a ./gradlew clean build
 # launch:
 
 ./gradlew uploadArchives -p kondor-core
@@ -12,3 +12,4 @@
 # select Staging Repositories, and close the corresponding one (empty desc is fine)1.
 # then click release and wait ~10 min to be able to download it
 # and then bouncing the version with SNAPSHOT in build.gradle
+# commit new shapshot version