+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 This list is not currently intended to be all-encompassing - it will document major and breaking API changes with their
 rationale when appropriate:
 
+### trunk
+
+Renamed JObject to ObjectNodeConverter and in JSealed typeFieldName to discriminatorFieldName and subtypesJObject to
+subConverters
+
 ### v1.2.0 - 4 Mar 2021
 
 Fixed test dependency (thanks to Asad Manji)